pyproject.toml
[project]
name = "flowguard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.2",
    "argon2-cffi>=25.1.0",
    "cryptography>=46.0.3",
    "fastapi[all]>=0.127.0",
    "google-genai>=1.56.0",
    "google-generativeai>=0.8.6",
    "passlib[argon2]>=1.7.4",
    "psycopg2>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.45",
]
